feat: 添加aboutAuthor组件; 滑动页面,目录固定
