+++ src/components/post/right/aboutAuthor.vue
@@ -1,0 +1,18 @@
+<template>
+    <div class="aboutAuthor">
+
+    </div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style scoped>
+.aboutAuthor {
+  background: #fff;
+  height: 200px;
+}
+</style>
